pyproject.toml
[tool.poetry]
name = "dify-eval"
version = "0.1.0"
description = ""
authors = ["libing <libing@healink.cn>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiohttp = "^3.10.6"
python-dotenv = "^1.0.1"
tenacity = "^8.5.0"
pandas = "^2.2.3"
langfuse = "^2.50.3"
pydantic = "^2.9.2"
ragas = "^0.1.20"
langchain_openai = "^0.1.0"
loguru = "^0.7.2"

[[tool.poetry.source]]
name = "aliyun-pypi"
url = "https://mirrors.aliyun.com/pypi/simple"
priority = "primary"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
